pyproject.toml
# pyproject.toml

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "djamplek8s"
authors = [{ name = "Maarten", email = "ikmaarten@gmail.com"}]
readme = "README.md"
requires-python = ">=3.8,<4"
dynamic = ["version", "description"]
license = { file = "LICENSE" }
keywords = ["django", "example", "k8s", "kubernetes", "project"]
classifiers = [
    "Development Status :: 2 - Pre-Alpha",
    "Environment :: Console",
    "Framework :: Django :: 3",
    "Framework :: Django",
    "License :: OSI Approved :: Apache Software License",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python",
    "Typing :: Typed",
]

dependencies = [
    "Django>=3.2,<4.0",
    "django-dotenv",
    "gunicorn",
    "psycopg2-binary",
]

[project.urls]
"Bug Tracker" = "https://github.com/maartenq/djamplek8s/issues"
"Source Code" = "https://github.com/maartenq/djamplek8s"
Documentation = "https://djamplek8s.readthedocs.io/en/latest/"

[project.optional-dependencies]
dev = [
    "black",
    "bump2version",
    "coverage[toml]",
    "flake8",
    "flit",
    "ipython",
    "mypy",
    "pre-commit",
    "requests",
    "tox",
]
docs = [
    "mkdocs",
]
test = [
    "coverage[toml]",
    "pre-commit",
    "pytest",
    "pytest-cov",
]

[project.scripts]
djamplek8s = "djamplek8s.manage:main"

[tool.black]
line-length = 79
experimental-string-processing = true

[tool.coverage.run]
branch = true

[tool.coverage.report]
exclude_lines = [
    "\\# pragma: no cover\\b",
    "^\\s*raise AssertionError\\b",
    "^\\s*raise NotImplementedError\\b",
    "^\\s*return NotImplemented\\b",
]
show_missing = true

[tool.isort]
line_length = 79
multi_line_output = 3
profile = "black"
skip_gitignore = true
